paginado corregido + filtrado por nombre
